+++ apps/web/src/app/components/product-switcher/product-switcher.component.html
@@ -1,45 +1,9 @@
 <ng-template [ngIf]="isEnabled">
   <button
     bitIconButton="bwi bwi-fw bwi-filter"
-    [bitMenuTriggerFor]="menuRef"
+    [bitMenuTriggerFor]="content?.menu"
     [buttonType]="buttonType"
     [attr.aria-label]="'switchProducts' | i18n"
   ></button>
-  <bit-menu #menuRef role="dialog">
-    <div class="tw-px-4 tw-py-2" *ngIf="products$ | async as products">
-      <!-- Bento options -->
-      <!-- grid-template-columns is dynamic so we can collapse empty columns  -->
-      <section
-        [ngStyle]="{
-          '--num-products': products.bento.length,
-          'grid-template-columns': 'repeat(min(var(--num-products,1),3),92px)'
-        }"
-        class="tw-grid tw-gap-2"
-      >
-        <a
-          *ngFor="let product of products['bento']"
-          [routerLink]="product.appRoute"
-          class="tw-flex tw-aspect-square tw-h-full tw-w-full tw-flex-col tw-items-center tw-justify-center tw-rounded tw-p-1 tw-outline-none hover:tw-bg-[rgba(26,65,172,0.03)] hover:tw-no-underline focus-visible:!tw-ring-2 focus-visible:!tw-ring-primary-500"
-          routerLinkActive="tw-font-bold tw-bg-primary-500 hover:tw-bg-primary-500 [&>*]:tw-text-contrast"
-          ariaCurrentWhenActive="page"
-        >
-          <i class="bwi {{ product.icon }} tw-text-4xl !tw-m-0 !tw-mb-1"></i>
-          <span class="tw-text-center tw-text-sm tw-leading-snug">{{ product.name }}</span>
-        </a>
-      </section>
-
-      <!-- Other options -->
-      <section
-        *ngIf="products.other.length > 0"
-        class="tw-mt-4 tw-flex tw-w-full tw-flex-col tw-border-0 tw-border-t tw-border-solid tw-border-t-text-muted tw-p-2 tw-pb-0"
-      >
-        <span class="tw-mb-1 tw-text-xs tw-text-muted">{{ "moreFromBitwarden" | i18n }}</span>
-        <a *ngFor="let product of products.other" [href]="product.marketingRoute">
-          <span class="tw-text-sm tw-text-primary-500"
-            ><i class="bwi {{ product.icon }} tw-m-0 !tw-mr-3"></i>{{ product.name }}</span
-          >
-        </a>
-      </section>
-    </div>
-  </bit-menu>
+  <product-switcher-content #content></product-switcher-content>
 </ng-template>
