+++ apps/web/src/app/components/product-switcher/product-switcher.module.ts
@@ -5,11 +5,12 @@ import { I18nPipe } from "@bitwarden/angular/pipes/i18n.pipe";
 
 import { SharedModule } from "../../shared";
 
+import { ProductSwitcherContentComponent } from "./product-switcher-content.component";
 import { ProductSwitcherComponent } from "./product-switcher.component";
 
 @NgModule({
   imports: [SharedModule, RouterModule],
-  declarations: [ProductSwitcherComponent],
+  declarations: [ProductSwitcherComponent, ProductSwitcherContentComponent],
   exports: [ProductSwitcherComponent],
   providers: [I18nPipe],
 })
